requirements.txt
# requirements.txt

absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crashtest==0.4.1
cvzone==1.6.1
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
easyocr==1.7.1
fastjsonschema==2.19.1
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.2
h11==0.14.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
importlib_metadata==7.1.0
imutils==0.5.4
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.3
keyring==24.3.1
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post5
pymongo==4.6.3
pyparsing==3.1.2
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-magic==0.4.27
python-socketio==5.11.2
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.8.1
requests==2.31.0
requests-toolbelt==1.0.0
scikit-image==0.23.1
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2024.4.18
tomli==2.0.1
tomlkit==0.12.4
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
trove-classifiers==2024.4.10
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.2
urllib3==2.2.1
virtualenv==20.25.3
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
